Adds Python/Flask REST Server That Proxies To Thrift Server
This directory contains a project that, assuming someone has
followed the instructions to run an HBase Thrift server,
can provision a fresh GCE instance with the required libraries
and then conncet to that Thrift server. To demonstrate
how the code works, it runs a local Flask server that exposes
a REST interface that proxies the CRUD calls to the Thrift
client.