Add Spark Scala Dataproc Wordcount Example
This is an update of the Spark example previously in this repo. Now includes instructions for packaging and running on local Spark as well as on Cloud Dataproc.

Currently has a test that runs with local Spark and deployed Bigtable. TODO is add full e2e test.

sbt seems more popular in spark world but was having problems with it so using Maven for now, TODO to switch.